refactor(db): rename file, create `IProfile`
